requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
dj-config-url==0.1.1
Django==4.0.1
django-environ==0.10.0
django-filter==23.2
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-filters
idna==3.4
oauthlib==3.2.2
phonenumbers==8.13.13
Pillow==9.5.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
dataclasses
typing-extensions
aiogram==2.25.1
marshmallow==3.20.1
python-telegram-bot==20.4
pydantic_core==2.3.0
marshmallow-dataclass==8.5.14
pytest==7.4.0
factory-boy==3.3.0








